build.sh: add meta blacklist feature
By adding a meta to the 'sources/blacklist' file, this meta will
not be added to build/conf/bblayers.conf and thus will not
be used in yocto build.
This is useful to keep a meta as a reference but disabled.
This is also useful to test a new meta which contains same .bbclass
or .inc files which relies on bblayers declaration order.
To blacklist a meta, just add its name in sources/blacklist,
one name per line.

Signed-off-by: Gilles DOFFE <gilles.doffe@savoirfairelinux.com>