tools:get-pkg-list-csv: fix image name
The image name of the image a package is installed with is extracted
from its manifest name. However, the current extraction method does not
remove the path of the file from its name. This makes sed fails as its
separator character is `/`, which is also found in file path.

Thus, filter the path out of an image manifest name when extracting an
image name.

Issue: #5514
Signed-off-by: Philip-Dylan Gleonec <philip-dylan.gleonec@savoirfairelinux.com>
Change-Id: Ic654c7e040be4953e1a9e4702c3e1b4e0c942b41