build.sh: refine regexp for excluding layers based on blacklist
Some layers can share a prefix in their naming, making the regexp
match in cases where layers were not blacklisted.

For instance, a sub-layer included in a layer collection which is named
"meta" will exclude all other sub-layers having "meta" in their name,
which results in blacklisting the complete collection.

Fix that issue by adding requirements on having the exact name to
consider exclusion of the layer.

Issue: #5905
Change-Id: I7e15a1d0b355987421c11ebe906d79fc2bc1bec2
Signed-off-by: Sébastien LE STUM <sebastien.le-stum@savoirfairelinux.com>